refactor: 유효성 검사 파일명과 메서드를 직관적으로 수정
